Turn the `Fit…` wrapper views into `ViewModifier`s
I don't know why this wasn't done like this to begin with, as `ViewModifier` is clearly the way to go.
The wrapper variants an now soft-deprecated (which means that they won't be proposed by Xcode when autocompleting, but they won't show warnings).
`WithLayoutMargins` is not concerned and is not soft-deprecated.